+++ templates/parts/component/link-more-product.php
@@ -6,7 +6,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0.0
- * @version  1.0.0
+ * @version  1.5.0
  */
 
 
@@ -16,9 +16,9 @@
 // Requirements check
 
 	if (
-			! is_search()
-			|| is_post_type_archive( 'product' )
-		) {
+		! is_search()
+		|| is_post_type_archive( 'product' )
+	) {
 		return;
 	}
 
@@ -26,14 +26,12 @@
 ?>
 
 <div class="link-more">
-	<a href="<?php echo esc_url( get_permalink( get_the_ID() ) ); ?>" class="button">
-		<?php
+	<a href="<?php the_permalink(); ?>" class="button"><?php
 
 		printf(
-				esc_html_x( 'View product%s&hellip;', '%s: Product name.', 'reykjavik' ),
-				the_title( '<span class="screen-reader-text"> &ldquo;', '&rdquo;</span>', false )
-			);
+			esc_html_x( 'View product%s&hellip;', '%s: Product name.', 'reykjavik' ),
+			the_title( '<span class="screen-reader-text"> &ldquo;', '&rdquo;</span>', false )
+		);
 
-		?>
-	</a>
+	?></a>
 </div>