+++ templates/parts/meta/entry-meta-element-date.php
@@ -6,16 +6,11 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0.0
- * @version  1.0.0
+ * @version  1.5.0
  */
 
 
 
-
-
-// Helper variables
-
-	$post_id = get_the_ID();
 
 
 ?>
@@ -24,7 +19,7 @@
 	<span class="entry-meta-description label-published">
 		<?php echo esc_html_x( 'Posted on:', 'Post meta info description: publish date.', 'reykjavik' ); ?>
 	</span>
-	<a href="<?php echo esc_url( get_permalink( $post_id ) ); ?>" rel="bookmark">
+	<a href="<?php the_permalink(); ?>" rel="bookmark">
 		<time datetime="<?php echo esc_attr( get_the_date( 'c' ) ); ?>" class="published" title="<?php echo esc_attr_x( 'Posted on:', 'Post meta info description: publish date.', 'reykjavik' ); echo ' ' . esc_attr( get_the_date() ); ?>">
 			<?php echo esc_html( get_the_date() ); ?>
 		</time>